requirements.txt
fastapi
uvicorn[standard]
playwright
python-dotenv
sqlalchemy
mysql-connector-python
pydantic-settings
httpx
pypdf
python-docx
